Add Jekyll configuration file
Set the name of the project and the version of markdown we will use.
Plus, modify the directories location of some default Jekyll
directories. Finally, add the Sass configuration in order to find where
find the *.scss files.

* Add Bundle Gemfile

This file is specific to Bundle (Ruby Gem) which is required to install
several Ruby Gems for both Bundle and Jekyll.