Add Header layout (#1)
* Rename the licence file

Adding the CC0 extension.

* Reorganise the project tree

Move the images, javascripts, and stylesheets directories into the
static directory.

* Add Jekyll configuration file

Set the name of the project and the version of markdown we will use.
Plus, modify the directories location of some default Jekyll
directories. Finally, add the Sass configuration in order to find where
find the *.scss files.

* Add Bundle Gemfile

This file is specific to Bundle (Ruby Gem) which is required to install
several Ruby Gems for both Bundle and Jekyll.

* Add .gitignore

Ignore backup files and some files or directories generated by the Ruby
Gems.

* Add header element

For the moment, there are a blank page for the Projects and
Documentations sections.