Implement Projects section
Each projects are represented by a card with a name and a version tag.
These projects are sorted by categories. A category is either a
programming language or a framework.

Note also the projects present in the Projects section are not clickable
for now, since they are not yet under the Stòiridh Project GitHub
repository.